Offer feed grouping
Feeds can now be organized in groups. The users can create feed groups and add feeds into them. The groups are shown as global drawer actions where the user can opt to read the feeds belonging in specific groups.
The users can also set the title of a feed to a different title from the one set by the feed provider.