Fix Sidebar Text Being Inappropriately Shown on Startup when Sidebar is Collapsed
When you closed alligator with the sidebar closed, it will remember that state the next time you launch it. However, the "all feeds", "settings", "manage feeds", and "about" sections seem to be in the uncollapsed state. This merge requests aims to fix that. One thing I wish I knew how to do is to make the text smoothly disappear like you have done for the feeds sections. Any instruction on how to do that and I would be glad to add it. I am not exactly well versed in qml/kirigami.