Fix open link with relativ urls
If a page https://example.org contains html like <a href="/link-to-
other-post-on-this-site">click here</a> QDesktopServices::openUrl tries
to open "/link-to-other-post-on-this-site" as a local file.

This patch adds the base url and the scheme from the page link for
relative urls to make them work.


(cherry picked from commit a1161bc6bae0f050c53c58454edd2cf6817c88e5)

Co-authored-by: Julius Künzel <julius.kuenzel@kde.org>