Add action to copy the URL
On my tablet, when I want to share a link with someone, currently I must open
in the browser, and then in the browser I must copy the URL and then finally
I can go to the chat and paste the link.

I think this is easier and I do it at least weekly.